Crate eq_balances

Equilibrium Balances Pallet

Equilibrium's Balances Pallet is a Substrate module that stores and modifies users balances. Apart from basic balances functionality this pallet was developed to support negative balance values.

Re-exports

pub use eq_primitives::signed_balance::SignedBalance::*;

pub use imbalances::NegativeImbalance;

pub use imbalances::PositiveImbalance;

Modules

balance_adapter
Module adapting eq_balances::EqCurrency to work as BasicCurrency in Substrate runtime

currency
Module to set list of available currencies and implement functions for currency operations

imbalances

Structs

Account
Pallet storage - balances for all accounts

GenesisConfig
Genesis config for the module, allow to build genesis storage.

Module

Enums

Call
Dispatchable calls.

Error
Balances pallet errors

RawEvent
Events for this module.

SignedBalance
Balance representation based on number generic Balance

Traits

BalanceChecker
Interface for balance checks

BalanceGetter
Balances reading interface

EqCurrency
An extension to Substrate standard Currency, adapting it to work in Equilibrium substrate

Trait
Substrate pallet configuration trait

WeightInfo

Type Definitions

Event
RawEvent specialized for the configuration Trait